Massive refactorisation of the BESSER WME Backend for better management, and update
