+++ besser/utilities/web_modeling_editor/backend/services/converters/buml_to_json/__init__.py
@@ -1,0 +1,20 @@
+"""
+BUML to JSON conversion modules.
+"""
+
+from .class_diagram_converter import class_buml_to_json, parse_buml_content
+from .state_machine_converter import state_machine_to_json
+from .agent_diagram_converter import agent_buml_to_json, analyze_function_node
+from .object_diagram_converter import object_buml_to_json
+from .project_converter import project_to_json, empty_model
+
+__all__ = [
+    'class_buml_to_json',
+    'parse_buml_content',
+    'state_machine_to_json',
+    'agent_buml_to_json',
+    'analyze_function_node',
+    'object_buml_to_json',
+    'project_to_json',
+    'empty_model'
+]
