+++ besser/utilities/web_modeling_editor/backend/services/converters/json_to_buml/__init__.py
@@ -1,0 +1,17 @@
+"""
+JSON to BUML conversion modules.
+"""
+
+from .class_diagram_processor import process_class_diagram
+from .object_diagram_processor import process_object_diagram
+from .state_machine_processor import process_state_machine
+from .agent_diagram_processor import process_agent_diagram
+from .project_converter import json_to_buml_project
+
+__all__ = [
+    'process_class_diagram',
+    'process_object_diagram',
+    'process_state_machine',
+    'process_agent_diagram',
+    'json_to_buml_project'
+]
