+++ besser/utilities/web_modeling_editor/backend/services/converters/parsers/__init__.py
@@ -1,0 +1,17 @@
+"""
+Parsers for converting between different formats.
+"""
+
+from .attribute_parser import parse_attribute
+from .method_parser import parse_method
+from .multiplicity_parser import parse_multiplicity
+from .ocl_parser import process_ocl_constraints
+from .text_parser import sanitize_text
+
+__all__ = [
+    'parse_attribute',
+    'parse_method', 
+    'parse_multiplicity',
+    'process_ocl_constraints',
+    'sanitize_text'
+]
