+++ besser/utilities/web_modeling_editor/backend/services/deployment/__init__.py
@@ -1,0 +1,9 @@
+"""
+Deployment module for handling application deployment.
+"""
+
+from .docker_deployment import run_docker_compose
+
+__all__ = [
+    "run_docker_compose",
+]
