+++ besser/utilities/web_modeling_editor/backend/services/utils/__init__.py
@@ -1,0 +1,19 @@
+"""
+Utils module for utility functions.
+"""
+
+from .layout_calculator import (
+    calculate_center_point,
+    determine_connection_direction,
+    calculate_connection_points,
+    calculate_path_points,
+    calculate_relationship_bounds,
+)
+
+__all__ = [
+    "calculate_center_point",
+    "determine_connection_direction",
+    "calculate_connection_points",
+    "calculate_path_points",
+    "calculate_relationship_bounds",
+]
