+++ besser/utilities/web_modeling_editor/backend/services/utils/layout_calculator.py
