+++ besser/utilities/web_modeling_editor/backend/services/validators/__init__.py
@@ -1,0 +1,9 @@
+"""
+Validators module for OCL checking and model validation.
+"""
+
+from .ocl_checker import check_ocl_constraint
+
+__all__ = [
+    "check_ocl_constraint",
+]
